+++ saleor/graphql/payment/mutations/payment/payment_check_balance.py
@@ -1,0 +1,111 @@
+import graphene
+from django.core.exceptions import ValidationError
+
+from .....payment import PaymentError
+from .....payment.error_codes import PaymentErrorCode
+from .....payment.utils import is_currency_supported
+from ....channel.utils import validate_channel
+from ....core import ResolveInfo
+from ....core.doc_category import DOC_CATEGORY_PAYMENTS
+from ....core.fields import JSONString
+from ....core.mutations import BaseMutation
+from ....core.scalars import PositiveDecimal
+from ....core.types import BaseInputObjectType
+from ....core.types import common as common_types
+from ....plugins.dataloaders import get_plugin_manager_promise
+
+
+class MoneyInput(graphene.InputObjectType):
+    currency = graphene.String(description="Currency code.", required=True)
+    amount = PositiveDecimal(description="Amount of money.", required=True)
+
+
+class CardInput(graphene.InputObjectType):
+    code = graphene.String(
+        description=(
+            "Payment method nonce, a token returned "
+            "by the appropriate provider's SDK."
+        ),
+        required=True,
+    )
+    cvc = graphene.String(description="Card security code.", required=False)
+    money = MoneyInput(
+        description="Information about currency and amount.", required=True
+    )
+
+
+class PaymentCheckBalanceInput(BaseInputObjectType):
+    gateway_id = graphene.types.String(
+        description="An ID of a payment gateway to check.", required=True
+    )
+    method = graphene.types.String(description="Payment method name.", required=True)
+    channel = graphene.String(
+        description="Slug of a channel for which the data should be returned.",
+        required=True,
+    )
+    card = CardInput(description="Information about card.", required=True)
+
+    class Meta:
+        doc_category = DOC_CATEGORY_PAYMENTS
+
+
+class PaymentCheckBalance(BaseMutation):
+    data = JSONString(description="Response from the gateway.")
+
+    class Arguments:
+        input = PaymentCheckBalanceInput(
+            description="Fields required to check payment balance.", required=True
+        )
+
+    class Meta:
+        description = "Check payment balance."
+        doc_category = DOC_CATEGORY_PAYMENTS
+        error_type_class = common_types.PaymentError
+        error_type_field = "payment_errors"
+
+    @classmethod
+    def perform_mutation(cls, _root, info: ResolveInfo, /, **data):
+        manager = get_plugin_manager_promise(info.context).get()
+        gateway_id = data["input"]["gateway_id"]
+        money = data["input"]["card"].get("money", {})
+
+        cls.validate_gateway(gateway_id, manager)
+        cls.validate_currency(money.currency, gateway_id, manager)
+
+        channel = data["input"].pop("channel")
+        validate_channel(channel, PaymentErrorCode)
+
+        try:
+            data = manager.check_payment_balance(data["input"], channel)
+        except PaymentError as e:
+            raise ValidationError(
+                str(e), code=PaymentErrorCode.BALANCE_CHECK_ERROR.value
+            )
+
+        return PaymentCheckBalance(data=data)
+
+    @classmethod
+    def validate_gateway(cls, gateway_id, manager):
+        gateways_id = [gateway.id for gateway in manager.list_payment_gateways()]
+
+        if gateway_id not in gateways_id:
+            raise ValidationError(
+                {
+                    "gateway_id": ValidationError(
+                        f"The gateway_id {gateway_id} is not available.",
+                        code=PaymentErrorCode.NOT_SUPPORTED_GATEWAY.value,
+                    )
+                }
+            )
+
+    @classmethod
+    def validate_currency(cls, currency, gateway_id, manager):
+        if not is_currency_supported(currency, gateway_id, manager):
+            raise ValidationError(
+                {
+                    "currency": ValidationError(
+                        f"The currency {currency} is not available for {gateway_id}.",
+                        code=PaymentErrorCode.NOT_SUPPORTED_GATEWAY.value,
+                    )
+                }
+            )
